FLINK-4798] CEPITCase.testSimpleKeyedPatternCEP test failure - Use TestListResultSink instead of writing tests results on disk to avoid IO issues - Move TestListResultSink from flink-tests_2.10 to flink-test-utils_2.10 to avoid IO issues unnecessary dependencies
