+++ flink-test-utils-parent/flink-test-utils/src/main/java/org/apache/flink/streaming/util/TestListWrapper.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.test.streaming.runtime.util;
+package org.apache.flink.streaming.util;
 
 import java.util.ArrayList;
 import java.util.Collections;
