+++ flink-tests/src/test/java/org/apache/flink/test/streaming/runtime/PartitionerITCase.java
@@ -28,7 +28,7 @@ import org.apache.flink.streaming.api.datastream.DataStream;
 import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
 import org.apache.flink.streaming.util.StreamingMultipleProgramsTestBase;
 import org.apache.flink.test.streaming.runtime.util.NoOpIntMap;
-import org.apache.flink.test.streaming.runtime.util.TestListResultSink;
+import org.apache.flink.streaming.util.TestListResultSink;
 
 import org.junit.Test;
